requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.6
-e git+https://github.com/gbozee/django-paystack.git@cfdf950e9f7fdc10fe6d4118b3a3327a68db68c9#egg=django_paystack
djangorestframework==3.11.0
future==0.18.2
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.9
Paperboy==1.0.1
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-async==0.6.2
rfc3986==1.4.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
